requirements.txt
python-pptx
openai
python-dotenv
requests 